feature: add strict enum parsing and access control improvements
* Rename NSObject `object` factory method to `instance` (old name kept for compatibility)
* Add strict enum/options parsing with validation output
* New access control methods: `mulleGainAccessWithUniquingSetIfAbsent` and `mulleRelinquishAccessWithUniquingSetIfAbsent`
* Add `MulleObjCGainAccessToObjects` and `MulleObjCRelinquishAccessToObjects` utility functions
* New `mulle_atomic_pointer_t` string setter for thread-safe ivar updates
* Improved NSInvocation error handling and alignment fixes
* Add `_MulleObjectForward` alias for dynamic forwarding
* Add methods to NSAutoreleasePool for single object release tracing
* removed `long double` support by default

Version bumped to 0.27. Added optional conditional compilation for long double type support. Reorganized NSObject category declarations for better API organization.

+++ src/class/NSMethodSignature.m
@@ -384,7 +384,7 @@ static MulleObjCMethodSignatureTypeInfo  *get_infos( NSMethodSignature *self)
 
    info        = &get_infos( self)[ 0];
    frame_size += info->natural_size;     // methodReturnLength
-   frame_size += alignof( long double);  // for alignment
+   frame_size += alignof( double);  // for alignment
 
    return( frame_size);
 }
